Added gigapixel-storymap widget and demo functionality
