+++ djgentelella/urls.py
@@ -61,6 +61,7 @@ for app in settings.INSTALLED_APPS:
     import_module_app_gt(app, 'gtcharts')
     import_module_app_gt(app, 'gtselects')
     import_module_app_gt(app, 'gttimeline')
+    import_module_app_gt(app, 'gtstorymap')
 
 base_urlpatterns = [
     url('gtapis/', include(routes.urls)),
